Implement `GetRawNode` local request msg and its handler
I've been trying to find a better way than exposing a static reference to RawNode using that function, but unfortunately, I haven't found one yet.

I would like to find a way to expose the APIs of RawNode in a sensible manner without relying on this particular API if possible.